refactor: Update mypy.ini to include mypy_drf_plugin
This commit updates the mypy.ini file to include the mypy_drf_plugin as a plugin. This plugin is necessary for type checking Django Rest Framework code. Additionally, the `disallow_untyped_defs` option is set to `false` to allow for untyped definitions. This change was made to address issues with incomplete type hints in the codebase.