refactor: Add API tests for creating posts and comments
Added API tests for creating posts and comments. The tests ensure that the API endpoints for creating posts and comments are functioning correctly. The tests include checking the response status code, as well as the correctness of the created post and comment data.

Refactored the `api/tests.py` file to include a new test case `PostAPITest` with a `test_create_post` method. This method uses the `reverse` function from `django.urls` to get the URL for creating a new post. It then sends a POST request to the URL with test data and asserts that the response status code is 201 (Created) and that the created post data matches the test data.

Also added a new file `api/tests/test_blog.py` with two test methods: `test_create_post` and `test_create_comment`. These methods use the `APIClient` class from `rest_framework.test` to authenticate the test user and send POST requests to the respective API endpoints for creating posts and comments. The methods assert that the response status code is 201 (Created) and that the created post and comment data match the test data.

The changes were made to improve the test coverage and ensure the correctness of the API functionality.